- /members/id-check  loginId를 받기위해 기존 @RequestBody와 loginDTO에서 @RequestParam을 이용해 String으로 받아 주었습니다.
loginId가 이미 테이블에 있을 시,
exceptionHandler를 이용한 처리로 변경했습니다.

- MemberRepository
MemberMapper와 역할이 중복된다 여겨 MemberRepository 클래스 및 인터페이스를 삭제했습니다.

- /members/join
/members/sign-up으로 url변경했습니다.

- MemberMapper.xml
기존 Optional을 이용해 Member 객체를 받던것에서
SQL을 수정해 select exists문으로 boolean 타입을 리턴해 좀더 성능을 높였습니다.

- MemberService.addMember()
롬복을 이용해 builder 패턴을 적용하여 좀 더 가시적으로 코드를 수정했습니다.

- 파일 수정
기존 깃에 대한 이해 부족으로인한
파일 중복 문제 등을 해결하였습니다.
