--- movieservice/src/main/java/study/movieservice/domain/member/Grade.java
@@ -1,6 +1,0 @@
-package study.movieservice.domain.member;
-
-public enum Grade {
-    BASIC,
-    VIP
-}
