--- movieservice/src/main/java/study/movieservice/domain/member/LoginIdDTO.java
@@ -1,8 +1,0 @@
-package study.movieservice.domain.member;
-
-import lombok.Getter;
-
-@Getter
-public class LoginIdDTO {
-    private String loginId;
-}
