--- movieservice/src/main/java/study/movieservice/domain/member/MemberDTO.java
@@ -1,11 +1,0 @@
-package study.movieservice.domain.member;
-
-import lombok.Getter;
-
-@Getter
-public class MemberDTO {
-    private String email;
-    private String loginId;
-    private String loginPassword;
-    private String nickname;
-}
