--- movieservice/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-#?? ??? ??.
-
-spring.datasource.url =jdbc:mariadb://localhost:3303/movie_service
-spring.datasource.username=root
-spring.datasource.password=alsdn120
-
-#MyBatis in xml convenient resultType. ( domian ??? ?? ?? )
-mybatis.type-aliases-package=study.movieservice.domain
-#db and java convert snake-case, camel-case
-mybatis.configuration.map-underscore-to-camel-case=true
-#session life
-server.servlet.session.timeout=60
