--- movieservice/src/main/resources/study/movieservice/repository/mapper/MemberMapper.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="study.movieservice.repository.mapper.MemberMapper">
-    <insert id="save" useGeneratedKeys="true" keyProperty="memberId">
-        insert into members(email, login_id, login_password, nickname, grade)
-        values (#{email}, #{loginId}, #{loginPassword},#{nickname},#{grade})
-    </insert>
-
-    <select id="findByLoginId" resultType="LoginIdDTO">
-        select login_id
-        from members
-        where login_id = #{loginId}
-    </select>
-</mapper>
-
