+++ src/main/java/study/movieservice/domain/member/Grade.java
@@ -1,0 +1,6 @@
+package study.movieservice.domain.member;
+
+public enum Grade {
+    BASIC,
+    VIP
+}
