+++ src/main/java/study/movieservice/domain/member/MemberDTO.java
@@ -1,0 +1,16 @@
+package study.movieservice.domain.member;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class MemberDTO {
+    private final String email;
+    private final String loginId;
+    private final String loginPassword;
+    private final String nickname;
+}
