+++ src/main/java/study/movieservice/repository/MemberMapper.java
@@ -1,0 +1,12 @@
+package study.movieservice.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+import study.movieservice.domain.member.Member;
+
+
+
+@Mapper
+public interface MemberMapper {
+    void save(Member member);
+    boolean findByLoginId(String loginId);
+}
