+++ src/main/java/study/movieservice/service/MemberService.java
@@ -1,0 +1,42 @@
+package study.movieservice.service;
+
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.mindrot.jbcrypt.BCrypt;
+import org.springframework.stereotype.Service;
+import study.movieservice.domain.member.Grade;
+import study.movieservice.domain.member.Member;
+import study.movieservice.domain.member.MemberDTO;
+import study.movieservice.repository.MemberMapper;
+
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+
+    private final MemberMapper memberMapper;
+
+    public void addMember(MemberDTO memberDTO){
+        String encodingPassword = BCrypt.hashpw(memberDTO.getLoginPassword(), BCrypt.gensalt());
+        Member member = Member.builder()
+                .email(memberDTO.getEmail())
+                .loginId(memberDTO.getLoginId())
+                .loginPassword(encodingPassword)
+                .nickname(memberDTO.getNickname())
+                .grade(Grade.BASIC).build();
+        memberMapper.save(member);
+    }
+
+    public boolean checkLoginId(String loginId){
+        boolean findFlag  = memberMapper.findByLoginId(loginId);
+
+        if(findFlag){
+            return true;
+        }
+
+        return false;
+    }
+
+}
